requirements.txt
python-dotenv
rapidfuzz
paho-mqtt
anthropic
